+++ examples/mouse.rb
@@ -3,8 +3,7 @@
 require_relative "../lib/chamomile"
 
 class MouseTracker
-  include Chamomile::Model
-  include Chamomile::Commands
+  include Chamomile::Application
 
   def initialize
     @x = 0
@@ -14,19 +13,16 @@ class MouseTracker
     @events = 0
   end
 
-  def update(msg)
-    case msg
-    when Chamomile::KeyMsg
-      return quit if msg.key == "q" || msg.ctrl?
-    when Chamomile::MouseMsg
-      @x = msg.x
-      @y = msg.y
-      @button = msg.button.to_s
-      @action = msg.action.to_s
-      @events += 1
-    end
-    nil
-  end
+  on_key("q") { quit }
+  on_key(:ctrl_c) { quit }
+
+  on_mouse { |e|
+    @x = e.x
+    @y = e.y
+    @button = e.button.to_s
+    @action = e.action.to_s
+    @events += 1
+  }
 
   def view
     lines = []